Initial implementation of a wordle guesser
Identifies two subproblems:
1. filtering possible answers from wordlist
2. choosing a word to guess next

This implementation does a reasonable job filtering but it is not
optimal. Notably it could do a better job keeping track of when more
than 1 letter is required.

The implementation of choice on the other hand has a lot of room for
improvement.

Uses 6.32 guesses/word compared to 9.16 in the sample implementation
while taking <1s compared to 20s.