wip: beam search
I replaced the one that calculates average cost per guess since I have
come to distrust floats and the old implementation was buggy. Not sure
it was caused by floats but also not sure it was not. In particular
there was a fun problem where widening the beam could cause the result
to deteriorate.

I have not observed the problem with wider beam leading to worse
results. What I have observed is increasing the number of words in play
decreasing the savings from beam search. I see no contradiction in this
since perhaps the new distribution happens to be a better fit for a
greedy algorithm but it is disconcerting, what if the benefit goes away
when all words are in play...

Benchmark:
149 {1: 1, 2: 12, 3: 27, 4: 8, 5: 1, 6: 1} 0.03136727202218026
146 {1: 1, 2: 13, 3: 26, 4: 9, 5: 1} 2.9128893520100974